Added the promised airfoil_loft.py
This is the airfoil_loft_vase.py that is much better documented and simpler. The vase mode is removed because it breaks the airfoil shape but I hope someone wants to take up the challenge of making a version of this that uses vase mode. See Readme.md for mode documentation.